apply verify to drawing api test
